requirements.txt
streamlit
python-dotenv
requests
gspread
google-auth
google-auth-oauthlib
google-auth-httplib2
google-genai
openai
Pillow
cloudinary
